% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{Dire Bonjour}
\usage{
dire_bonjour(prenom = "toi")
}
\arguments{
\item{prenom}{chaine de caracteres}
}
\value{
renvoie un message qui dit bonjour a quelqu'un
}
\description{
Dire Bonjour
}
\examples{
dire_bonjour('Arnaud')
dire_bonjour()
}
